[dv] Fix some signed/unsigned comparison warnings
It seems that Xcelium builds stuff with -Wsign-compare by default.
Fixing these up seems safe enough and avoids a bit of unnecessary
verbosity in the logs.

+++ hw/ip/otbn/dv/model/iss_wrapper.cc
@@ -471,7 +471,7 @@ uint32_t ISSWrapper::step_crc(const std::array<uint8_t, 6> &item,
 
   std::ostringstream oss;
   oss << std::hex << "step_crc 0x" << std::setfill('0');
-  for (int i = 0; i < item.size(); ++i) {
+  for (size_t i = 0; i < item.size(); ++i) {
     oss << std::setw(2) << (int)item[5 - i];
   }
   oss << " 0x" << std::setw(8) << state << "\n";
